pyproject.toml
[project]
name = "rag-based-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.21",
    "llama-index-readers-file>=0.6.0",
    "openai>=2.34.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.17.1",
    "requests>=2.32.0",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.57.0",
    "uvicorn>=0.46.0",
]
